Add option --notify-absolute to show absolute volume in notification
If you call the script multiple times in a row (e.g. through volume
wheel) get multiple notifications for every change. These might get
stacked due to similar contents, but are not particular useful.

This implements to show the absolute volume of sink or sink input in the
notification message with the option `--notify-absolute`. To make it
possible to update an old/recent notifications, it is necessary to know
the id property of the notification. It will be written to a temporary
file (which is never cleaned up – fix?) and read if available to set the
id property of the notification before showing. This will overwrite the
last notification and enables changing notification.